Textfragment uint32_t

+++ Source/core/Portability.cpp
@@ -448,8 +448,8 @@ namespace Core {
     TextFragment ClassNameOnly(const char name[]) {
 
         TextFragment result(Demangled(name));
-        uint16_t index = 0;
-        uint16_t lastIndex = static_cast<uint16_t>(~0);
+        uint32_t index = 0;
+        uint32_t lastIndex = static_cast<uint32_t>(~0);
 
         while ((index < result.Length()) && (result[index] != '<')) {
             if (result[index] == ':') {
